More implementing of client-server interaction
First, I forgot to add the send for the runner, because either way the
other client wouldn’t ever receive anything. Second, I changed the
DataOutputStream to an ObjectOutputStream, just to see if it would work
or not. Lastly, I added a debug println statement to show if the
BufferedReader in the PacketSend was ready.