Futures work and game is playable...for two moves.
The future is really great for…well it works and there’s no extremely
code. However, this currently lasts for two moves, before both players
get to the “hitting turn” simultaneously. So that’s what I’m working on
next.

Also, I temporarily made the player # display on its board so it’s
easier for me to see if the game is working or not.